Dima&Jerry: delete angepasst & populate, Validations hinzugefügt

+++ v1/src/controller/events.controller.js
@@ -2,7 +2,8 @@ import { StatusCodes } from 'http-status-codes';
 import Event from '../models/Event.model.js';
 import { 
   BadRequestError,
-  NotFound
+  NotFound,
+  Unauthorized
 } from '../errorHandler/index.js';
 
 async function createEvent(req, res, next) {
@@ -24,22 +25,25 @@ async function getAllEvents(req,res,next) {
     try {
         const events = await Event.find();
 
+        if(!events){
+          throw new BadRequestError('Konnte keine Events finden!');
+        }
         res.status(StatusCodes.OK).json({ 
           msg: 'Erfolgreicher Empfang von Eventdaten',
           events 
         });
     } catch(error) {
-        next(new BadRequestError(error));
+        next(error);
     }
 };
 
 async function getSingleEvent(req,res,next) {
     const {id:eventId} = req.params;
     try{ 
-        const event = await Event.findOne({_id: eventId});
+        const event = await Event.findOne({ _id: eventId });
 
         if(!event){
-          throw new BadRequestError(`Kein passendes Event gefunden!`)
+          throw new BadRequestError(`Kein passendes Event gefunden!`);
         }
 
         await event.populate({ path: 'creator', select: 'username -_id'})
@@ -49,7 +53,7 @@ async function getSingleEvent(req,res,next) {
             event
         });
     } catch(error) {
-        next(new BadRequestError(error));
+        next(error);
     }
 };
 
@@ -74,7 +78,6 @@ async function updateOneEvent(req,res,next) {
             newEvent
         })
     } catch(error) {
-        // Internal Server Error
         next(error);
     }
 };
@@ -89,11 +92,16 @@ async function deleteOneEvent(req, res, next) {
     if(!event) {
       throw new NotFound(`Kein Event mit der ID: ${eventId} gefunden!`);
     }
+
+    if(event.creator.toString() !== req.user.userId){
+      throw new Unauthorized('Dieses Event darf nur der Ersteller löschen!')
+    }
+
     await Event.findOneAndDelete({ _id: eventId });
    
 
     res.status(StatusCodes.OK).json({
-      msg: `Event mit der ID: ${eventId} wurde gelöscht!`
+      msg: `Das Event ${event.title} wurde gelöscht!`
     });
   } catch (error) {
     next(error);
